Category by id

+++ backend/src/main/java/com/eblj/catalog/repositories/CategoryRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.eblj.catalog.entities.Category;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer>{/*JpaRepository é generico e espera um tipo quel o tipo do id*/
+public interface CategoryRepository extends JpaRepository<Category, Long>{/*JpaRepository é generico e espera um tipo quel o tipo do id*/
